feat: Add the optional ability to disable toolchain registration after install
We have a use case where the default toolchain + platform registration doesn't
work for us, so we would like to be able to inject our own. We can already do
this with the custom BUILD file option, but we also need to be able to
disable the generation of the toolchain + platform files that happen by
default.

This PR introduces a new option `register_toolchain`, which defaults to
`True`, and is plumbed down to the `install_and_setup_sdk` rule.

The default behavior is unchanged, unless users pass in their own value
for this flag.