Update to SDK versoin 5.0.10 from upstream and remove '-R' flag
Old versions of SDK was no longer available upstream for download.
Refresh to version 5.0.10.

FOr some reason there was an error during installation of SDK when
using '-R' flag. I can not find any reason or information why it
is used. Removed to succeed installation and building example
in tests/e2e.