Introduce clang and gcc
Create two different toolchains. One for gcc and one for clang.
The original one for gcc is named as before. New one with clang
is named 'cc-ttolchain-target-clang'.

One gotcha here is that version number of gcc and clang are
hardcoded in to config right now. gcc 13.3.0 and clag 20 is assumed
in the SDK.